refactor(workbench): merge Container definitions into component files
