+++ app/components/workbench/Workbench.tsx
@@ -29,7 +29,7 @@ import { defaultSidebarWidth } from '../../reducers/ui'
 import { Resizable } from '../Resizable'
 import Layout from '../Layout'
 // import UnlinkedDataset from './UnlinkedDataset'
-import DatasetComponent from '../DatasetComponent'
+import DatasetComponent from './DatasetComponent'
 import NoDatasetSelected from './NoDatasetSelected'
 import HeaderColumnButton from '../chrome/HeaderColumnButton'
 import Hamburger from '../chrome/Hamburger'
